+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## [2026-05-11]
 - feat(api): enforce non-optional fields in OpenAPI schema
 - feat: add users module with roles and leaves
+- chore(web): disable animals route in UI and via direct URL